// apps/web/src/app/(protected)/vms/interfaces.tsx
import { User } from 'next-auth'

export interface VirtualMachine {
  id?: {
    $olid?: string
  }
  uid: string
  metadata?: {
    name?: string
    generatename?: string
    namespace?: string
    selflink?: string
    uid?: string
    resourceversion?: string
    creationtimestamp?: {
      time?: {
        date?: {
          numberLong?: string
        }
      }
    }
    deletiontimestamp?: string | null
    labels?: string | null
    annotations?: string | null
    owerreferences?: string | null
    finalizers?: string[] | null
    managedfields?: string | null
  }
  rormeta: {
    version?: string
    lastreported?: string
    internal?: boolean
    hash?: string
    owerref?: {
      scope?: string
      subject?: string
    }
    action?: string
    tags?: string[] | null
  }
  typemeta?: {
    kind?: string
    apiversion?: string
  }
  virtualmachine?: {
    externalid?: string | null
    spec?: {
      cpu?: {
        sockets?: number
        corespersocket?: number
      }
      name?: string
      disks?: string | null
      memory?: {
        sizebytes?: number
      }
    }
    status?: {
      lastupdated?: {
        time?: {
          date?: string
        }
      }
      location?: string | null
      cpu?: {
        units?: number | null
        usage?: number
        resourcevirtualmachinecpuspec?: {
          sockets?: number
          corespersocket?: number
        }
      }
      tags?: {} | null
      state?: {
        state?: string | null
        reason?: string
        time?: string | undefined
      }
      disks?: [
        {
          usagebytes?: number
          ismounted?: boolean
          resourcevirtualmachinediskspec?: {
            id?: string
            name?: string
            type?: string
            sizebytes?: number
          }
        },
      ]
      memory?: {
        unit?: string
        usage?: number
        resourcevirtualmachinememoryspec?: {
          sizebytes?: number
        }
      }
      networks?:
        | [
            {
              id?: string
              dns?: string
              ipv4?: string
              ipv6?: string
              mask?: string
              gateway?: string
              mac?: string
            },
          ]
        | null
      operatingsystem?: {
        id?: string | null
        name?: string | null
        family?: string | null
        version?: string | null
        hostname?: string | null
        powerstate?: string | null
        toolversion?: string | null
        architecture?: string | null
      }
    }
    provider?: string | null
  }
}

export interface VmResponse {
  resources: VirtualMachine[]
}

export interface VMTableProps {
  metadata_name?: string
  os_id?: string | null
  powerstate?: string | null
}

export interface Params {
  view?: 'grid' | 'list'
  page?: number
  limit?: number
  sort?: string
  order?: 'asc' | 'desc'
  filters?: string
}

export interface PageViewProps {
  className?: string
  user: User
  vms: VirtualMachine[]
  params: Params
}